requirements.txt
--constraint https://raw.githubusercontent.com/apache/airflow/constraints-2.10.4/constraints-3.12.txt
apache-airflow==2.10.4
apache-airflow-providers-ssh
pydantic
pytest